+++ frontend/src/components/kl-add-item.js
@@ -1,5 +1,5 @@
 import * as Item from 'domain/Item';
-import {parseValidationErrors} from 'domain/validation';
+import {renderValidationErrors} from '../renderErrors';
 import { focus } from 'vue-focus';
 export default {
     directives: { focus: focus },
@@ -22,15 +22,15 @@ export default {
                 this.$emit('addItem', item);
             }).catch(err => {
                 if (err.status === 400) {
-                    let parsedErrors = parseValidationErrors(err.body.errors);
-                    parsedErrors.forEach(error => toastr.error(error));
+                    let renderedErrors = renderValidationErrors(err.body.errors);
+                    renderedErrors.forEach(error => toastr.error(error));
                 } else
                     toastr.error('Oops, something went wrong');
             });
         },
         statControl: function () {
-            if (this.minDmg < 2) this.minDmg = 2;
-            if (this.maxDmg <= this.minDmg) this.maxDmg = this.minDmg + 1;
+            //if (this.minDmg < 2) this.minDmg = 2;
+            //if (this.maxDmg <= this.minDmg) this.maxDmg = this.minDmg + 1;
         }
     },
     watch: {
